refactor: separate MetaInfo and Pagination type
