+++ src/models/MetaInfo.ts
@@ -10,15 +10,6 @@ export interface ExternalUrls {
   id?: string;
 }
 
-export interface Pagination {
-  href: string;
-  limit: number;
-  next?: string;
-  offset: number;
-  previous?: string | null;
-  total: number;
-}
-
 export interface MetaInfo {
   uri: string;
   id: string;
