+++ src/models/Pagination.ts
@@ -1,0 +1,8 @@
+export interface Pagination {
+  href: string;
+  limit: number;
+  next?: string;
+  offset: number;
+  previous?: string | null;
+  total: number;
+}
